[feat] 이용습관 진단지 결과 저장 구현

+++ src/main/java/com/coders/boardgame/repository/HabitSurveyRepository.java
@@ -3,5 +3,5 @@ package com.coders.boardgame.repository;
 import com.coders.boardgame.entity.HabitSurvey;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HabitSurveyRepository extends JpaRepository<HabitSurvey, Long> {
+public interface HabitSurveyRepository extends JpaRepository<HabitSurvey, Integer> {
 }
